generalized dropdown toggle
